requirements.txt
# Task Mining Multi-Agent System Requirements

# Core Data Processing
pandas>=1.5.0
numpy>=1.24.0

# Visualization
altair>=5.0.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0

# Gradio Frontend
gradio>=4.0.0

# Date/Time Processing
python-dateutil>=2.8.0

# PDF Generation
reportlab>=4.0.0

# Image Processing for PDF export
Pillow>=9.5.0

# HTTP Requests
requests>=2.31.0

# AI/ML Integration
openai>=1.0.0

# Development Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Optional: For enhanced chart rendering
selenium>=4.15.0
webdriver-manager>=4.0.0

# Optional: For advanced analytics
scipy>=1.11.0
scikit-learn>=1.3.0
